Extract a non-observable UseCase for quote cart Checkout
